Added cast to int to spinner value
git-svn-id: https://svn.code.sf.net/p/sweethome3d/code/trunk/SweetHome3D@716 03e323e1-6c6c-46a5-9f71-c4ce5d8c6204

+++ src/com/eteks/sweethome3d/swing/Home3DAttributesPanel.java
@@ -206,7 +206,7 @@ public class Home3DAttributesPanel extends JPanel {
    */
   private void updateComponents(Home home) {
     ObserverCamera observerCamera = home.getObserverCamera();
-    this.observerFieldOfViewSpinner.setValue(Math.round(Math.toDegrees(observerCamera.getFieldOfView())));
+    this.observerFieldOfViewSpinner.setValue((int)Math.round(Math.toDegrees(observerCamera.getFieldOfView())));
     ((NullableSpinner.NullableSpinnerLengthModel)this.observerHeightSpinner.getModel()).
         setLength((float)Math.round(observerCamera.getHeight() * 100) / 100);
     this.groundColorButton.setColor(home.getGroundColor());
